requirements.txt
Flask==3.0.0
opencv-python-headless==4.9.0.80
pytesseract==0.3.10
easyocr==1.7.1
Pillow==10.2.0
numpy==1.26.3
google-generativeai==0.3.2
gunicorn==21.2.0
